+++ main.py
@@ -1,6 +1,6 @@
-from model_loader import ModelLoader
-from animation import Animation
-from util import get_global_settings
+from animation.model_loader import ModelLoader
+from animation.animation import Animation
+from animation.util import get_global_settings
 
 if __name__ == "__main__":
     # load settings
